Version 0.0.3.136:
- Bugfix: Shell stürzt nicht mehr ab, wenn man den Schreibcursor nach links bewegt
- Bugfix: Kein #GPF mehr, wenn ein USB-Stick schon beim boot steckt (EHCI)
- Einige fehlende E0-Tasten ergänzt - Entf, Pos1, End, ...-Tasten funktionieren wieder (u.a. von der Shell unterstützt)
- Timer-Ticks jetzt als uint64_t gespeichert (Rechnungen weiterhin mit uint32_t, da keine 64-bit-Unterstützung)
- video_updateCursor in vga_updateCursor umbenannt. Einsatz überarbeitet.

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@1337 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/timer.c
@@ -11,7 +11,7 @@
 
 
 static uint16_t systemfrequency; // system frequency
-static volatile uint32_t timer_ticks = 0;
+static volatile uint64_t timer_ticks = 0;
 
 void timer_install(uint16_t sysfreq)
 {
@@ -23,11 +23,11 @@ void timer_install(uint16_t sysfreq)
 
 uint32_t timer_getSeconds()
 {
-    return(timer_ticks/systemfrequency);
+    return((uint32_t)timer_ticks/systemfrequency);
 }
 uint32_t timer_getMilliseconds()
 {
-    return((timer_ticks*1000)/systemfrequency);
+    return(((uint32_t)timer_ticks*1000)/systemfrequency);
 }
 uint64_t timer_getTicks()
 {
